# docs/swagger.yaml
basePath: /subscriptions
definitions:
  dto.CreateSubscriptionRequest:
    description: Request body for creating a new subscription
    properties:
      end_date:
        example: 12-2024
        type: string
      price:
        example: 999
        minimum: 1
        type: integer
      service_name:
        enum:
        - Netflix
        - Spotify
        - Apple Music
        example: Netflix
        maxLength: 100
        minLength: 1
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  dto.SubscriptionResponse:
    description: Response containing updated subscription fields
    properties:
      end_date:
        example: 12-2024
        type: string
      id:
        example: 1
        type: integer
      price:
        example: 999
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  dto.SumSubscriptionRequest:
    description: Request body for collecting subscription statistics
    properties:
      end_period:
        example: 12-2024
        type: string
      service_name:
        example: Netflix
        maxLength: 100
        minLength: 1
        type: string
      start_period:
        example: 01-2024
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - end_period
    - start_period
    type: object
  dto.SumSubscriptionResponse:
    description: Response containing subscription details
    properties:
      total_count:
        example: 10
        type: integer
      total_sum:
        example: 15
        type: integer
    type: object
  dto.UpdateSubscriptionRequest:
    description: Request body for updating an existing subscription
    properties:
      end_date:
        example: 01-2024
        type: string
      price:
        example: 999
        type: integer
      service_name:
        example: Netflix
        maxLength: 100
        minLength: 1
        type: string
    type: object
  dto.UpdateSubscriptionResponse:
    description: Response containing subscription statistics
    properties:
      end_date:
        example: 12-2024
        type: string
      id:
        example: 1
        type: integer
      price:
        example: 999
        type: integer
      service_name:
        example: Netflix
        type: string
    type: object
info:
  contact: {}
  description: CRUDL subscription service
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: List of subscriptions
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionResponse'
            type: array
      summary: List all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription with the provided details
      parameters:
      - description: Subscription creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Subscription created successfully
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Subscription deleted successfully
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Retrieve a subscription by its unique ID
      parameters:
      - description: Subscription ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Subscription found
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
      summary: Get a subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update an existing subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Subscription update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Subscription updated successfully
          schema:
            $ref: '#/definitions/dto.UpdateSubscriptionResponse'
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/stats:
    post:
      consumes:
      - application/json
      description: Collect statistics for subscriptions within a date range
      parameters:
      - description: Statistics collection request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SumSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Statistics collected successfully
          schema:
            $ref: '#/definitions/dto.SumSubscriptionResponse'
      summary: Collect subscription statistics
      tags:
      - statistics
swagger: "2.0"